requirements.txt
-f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
Flask_Cors==3.0.9
torch==1.8.1+cpu
transformers==4.6.1
pandas==1.3.0
Flask==1.1.2
numpy==1.21.1
scikit_learn==0.24.2
torchinfo==1.5.2
openpyxl==3.0.7
twython==3.9.1
ekphrasis==0.5.1
tweet-preprocessor==0.6.0
emot==3.1
emoji==1.4.2
imbalanced-learn==0.8.0